+++ DomainModels/Room.cs
@@ -11,7 +11,6 @@ namespace DomainModels
         available,
         underMaintenece,
         needsCleaning
-
     }
 
     public class Room
@@ -19,7 +18,6 @@ namespace DomainModels
         public int id { get; set; }
         public string type { get; set; }
         public int price { get; set; }
-
         public List<DateTime> bookedDays { get; set; } = new List<DateTime>();
 
         Status status { get; set; }
